Chnages with detail page and Type of pokemon
