+++ package.json
@@ -13,8 +13,13 @@
     "@testing-library/user-event": "^13.2.1",
     "react": "^18.2.0",
     "react-dom": "^18.2.0",
+    "react-query": "^3.39.2",
     "react-redux": "^8.0.2",
+    "react-router": "^6.4.2",
+    "react-router-dom": "^6.4.2",
     "react-scripts": "5.0.1",
+    "sass": "^1.55.0",
+    "scss": "^0.2.4",
     "web-vitals": "^2.1.0"
   },
   "scripts": {
