+++ src/App.js
@@ -1,12 +1,19 @@
 import "./App.css";
 import { DetailPage } from "./Components/PokeDetails/DetailPage";
 import { PokeCard } from "./Components/PokeDetails/PokeCard";
-import TopBar from "./Components/searchbar/SearchBar";
+import { BrowserRouter as Router, Route, Routes } from "react-router-dom";
 
 function App() {
   return (
     <div className="App">
-      <PokeCard />
+      {/* <PokeCard /> */}
+      <Router>
+        <Routes>
+          <Route path="/" element={<PokeCard />} />
+          <Route path="/pokedetails/" element={<DetailPage />} />
+          {/* <Route path="/pokedetails/:id" element={<DetailPage />} /> // works with query by passsing id */}
+        </Routes>
+      </Router>
     </div>
   );
 }
