+++ src/Components/PokeDetails/PokeCard.scss
@@ -1,0 +1,53 @@
+.card2 {
+  display: block;
+  position: relative;
+  max-width: 262px;
+  background-color: #f2f8f9;
+  border-radius: 4px;
+  text-decoration: none;
+  z-index: 0;
+  overflow: hidden;
+
+  &:before {
+    content: "";
+    position: absolute;
+    z-index: -1;
+    top: -26px;
+    right: -26px;
+    background: #e3f2fd;
+    height: 32px;
+    width: 32px;
+    border-radius: 32px;
+    transform: scale(1);
+    transform-origin: 50% 50%;
+    transition: transform 0.25s ease-out;
+  }
+
+  &:hover:before {
+    transform: scale(21);
+  }
+}
+
+.pokename {
+  opacity: 1;
+  transition: opacity .6s ease-in-out;
+  color: #6d6d6d;
+}
+
+.poketype {
+  padding: 4px;
+  color: #fff !important;
+  transition: background-color 1s ease-in, width 1.5s ease-in-out;
+  background-color: #81c784 !important;
+  border-radius: 10px;
+}
+
+.pokedetailsection {
+  position: relative;
+  padding-bottom: 3rem;
+  justify-content: center !important;
+  display: flex;
+  flex-wrap: wrap;
+  margin-right: -15px;
+  margin-left: -15px;
+}
