+++ src/Redux/reducers/PokeReducer.js
@@ -1,11 +1,8 @@
 import { PokeActionTypes } from "../constants/PokeActionTypes";
 
 export const initialState = {
-  Records: [
-    {
-      id: 1,
-    },
-  ],
+  Records: [],
+  selectedRecords: [],
 };
 
 const PokeReducer = (state = initialState, action) => {
@@ -13,9 +10,16 @@ const PokeReducer = (state = initialState, action) => {
 
   switch (action.type) {
     case PokeActionTypes.UPDATE_POKEMON:
-      newState.Records = action.data;
+      newState.Records = action.payload;
       return newState;
 
+    case PokeActionTypes.SELECTED_POKEMON:
+      newState.selectedRecords = action.payload;
+      // return {
+      //   ...state,
+      //   Records: action.payload,
+      // };
+      return newState;
     default:
       return state;
   }
